✨ (NamingHistory.jsx, AzureNamingContext.jsx, useAzureNaming.js): refactor naming history management by integrating local storage, streamline form state handling, and enhance configuration loading logic for improved performance and user experience.
